# apps/hub_api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PotStatePatch": {
        "properties": {
          "De_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "De Mm"
          },
          "Dr_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dr Mm"
          },
          "Kcb_struct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kcb Struct"
          },
          "Ke_prev": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ke Prev"
          },
          "REW_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rew Mm"
          },
          "c_AC": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "C Ac"
          },
          "c_aero": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "C Aero"
          },
          "last_irrigation_ts": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Irrigation Ts"
          },
          "tau_e_h": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tau E H"
          }
        },
        "title": "PotStatePatch",
        "type": "object"
      },
...
      "StepContext": {
        "description": "Snapshot of the inputs that produced a step result.",
        "properties": {
          "dt_h": {
            "title": "Dt H",
            "type": "number"
          },
          "pot_area_m2": {
            "title": "Pot Area M2",
            "type": "number"
          },
          "sensors": {
            "$ref": "#/components/schemas/StepSensors"
          }
        },
        "required": [
          "sensors",
          "dt_h",
          "pot_area_m2"
        ],
        "title": "StepContext",
        "type": "object"
      },
      "StepResult": {
        "properties": {
          "De_mm": {
            "title": "De Mm",
            "type": "number"
          },
          "Dr_mm": {
            "title": "Dr Mm",
            "type": "number"
          },
          "ET0_mm": {
            "title": "Et0 Mm",
            "type": "number"
          },
          "ETc_model_mm": {
            "title": "Etc Model Mm",
            "type": "number"
          },
          "ETc_obs_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Etc Obs Mm"
          },
          "Kcb_eff": {
            "title": "Kcb Eff",
            "type": "number"
          },
          "Kcb_struct": {
            "title": "Kcb Struct",
            "type": "number"
          },
          "Ke": {
            "title": "Ke",
            "type": "number"
          },
          "Ks": {
            "title": "Ks",
            "type": "number"
          },
          "REW_mm": {
            "title": "Rew Mm",
            "type": "number"
          },
          "c_aero": {
            "title": "C Aero",
            "type": "number"
          },
          "context": {
            "$ref": "#/components/schemas/StepContext"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "need_irrigation": {
            "title": "Need Irrigation",
            "type": "boolean"
          },
          "recommend_mm": {
            "title": "Recommend Mm",
            "type": "number"
          },
          "tau_e_h": {
            "title": "Tau E H",
            "type": "number"
          }
        },
        "required": [
          "ET0_mm",
          "ETc_model_mm",
          "ETc_obs_mm",
          "Kcb_struct",
          "Kcb_eff",
          "c_aero",
          "Ke",
          "Ks",
          "De_mm",
          "Dr_mm",
          "REW_mm",
          "tau_e_h",
          "need_irrigation",
          "recommend_mm",
          "context"
        ],
        "title": "StepResult",
        "type": "object"
      },
      "StepSensors": {
        "description": "Sensor inputs collected over a control interval.",
        "properties": {
          "AC_on": {
            "default": false,
            "title": "Ac On",
            "type": "boolean"
          },
          "RH_pct": {
            "title": "Rh Pct",
            "type": "number"
          },
          "Rs_MJ_m2_h": {
            "title": "Rs Mj M2 H",
            "type": "number"
          },
          "T_C": {
            "title": "T C",
            "type": "number"
          },
          "dStorage_mL": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dstorage Ml"
          },
          "drain_mL": {
            "default": 0.0,
            "title": "Drain Ml",
            "type": "number"
          },
          "inflow_mL": {
            "default": 0.0,
            "title": "Inflow Ml",
            "type": "number"
          },
          "theta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theta"
          },
          "u2_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "U2 Ms"
          }
        },
        "required": [
          "T_C",
          "RH_pct",
          "Rs_MJ_m2_h"
        ],
        "title": "StepSensors",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ProjectPlant Hub",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/etkc/state/{plant_id}": {
      "get": {
        "operationId": "api_get_state_api_v1_etkc_state__plant_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "plant_id",
            "required": true,
            "schema": {
              "title": "Plant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Get State Api V1 Etkc State  Plant Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Get State",
        "tags": [
          "etkc"
        ]
      },
      "put": {
        "operationId": "api_put_state_api_v1_etkc_state__plant_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "plant_id",
            "required": true,
            "schema": {
              "title": "Plant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PotStatePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Put State Api V1 Etkc State  Plant Id  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Put State",
        "tags": [
          "etkc"
        ]
      }
    },
    "/api/v1/etkc/step/{plant_id}": {
      "post": {
        "operationId": "api_step_api_v1_etkc_step__plant_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "plant_id",
            "required": true,
            "schema": {
              "title": "Plant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepSensors"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Step",
        "tags": [
          "etkc"
        ]
      }
    },
...
  }
}